Consolidate sample songs into single file
- Remove dev-mode-sample-songs.ts file entirely
- Move populateDevModeSampleSongs function into sample-songs.ts
- Simplify code architecture with single responsibility
- Reduce unnecessary file complexity and imports
- All tests still pass (374 passed)

This eliminates the confusing two-file pattern and creates a cleaner,
more maintainable solution for sample songs handling.
